+++ CHANGELOG.md
@@ -8,6 +8,8 @@ glance - what to expact from upgrading to a new version.
 
 ### Added
 
+- `!leave` command which allows leaving a Discord guild.
+
 ### Changed
 
 ### Fixed